requirements.txt
# Bot framework — поддержка Bot API 9.6 (sendMessageDraft и др.)
aiogram==3.27.0

# Web сервер для webhook
aiohttp>=3.10,<3.13

# Хранилище FSM и кэш
redis>=5.0

# Конфиги из .env
environs>=11.0

# OpenAI совместимый клиент (GPT, Groq и др.)
openai>=1.51

# HTTP клиент с поддержкой proxy
httpx>=0.27

# Async I/O для файлов
aiofiles>=24.0

# Парсинг документов
python-docx>=1.1
pypdf>=5.0           # современная замена устаревшему PyPDF2

# Изображения
Pillow>=10.0

# Опционально: ускоряет asyncio на Linux в 2-4 раза
uvloop>=0.20; sys_platform != "win32"
